+++ jax/numpy/__init__.py
@@ -22,9 +22,6 @@ from jax._src.basearray import Array as ndarray
 
 from jax._src.numpy.lax_numpy import (
     ComplexWarning as ComplexWarning,
-    NINF as NINF,  # TODO: removed in Numpy 1.26
-    NZERO as NZERO,  # TODO: removed in Numpy 1.26
-    PZERO as PZERO,  # TODO: removed in Numpy 1.26
     allclose as allclose,
     angle as angle,
     append as append,
@@ -443,6 +440,19 @@ _deprecations = {
     "sometrue": (
         "jax.numpy.sometrue is deprecated. Use jax.numpy.any",
         any,
+    ),
+    # Added August 10, 2023:
+    "NINF": (
+        "jax.numpy.NINF is deprecated. Use -jax.numpy.inf instead.",
+        -inf,
+    ),
+    "NZERO": (
+        "jax.numpy.NZERO is deprecated. Use -0.0 instead.",
+        -0.0,
+    ),
+    "PZERO": (
+        "jax.numpy.PZERO is deprecated. Use 0.0 instead.",
+        0.0,
     ),
 }